Minimal fix to input file reading
body contents are not cleared when reading a new file. Technically this should be fixed in Body.hh, Mesh.hh, PField.hh, etc, but it's not worth the effort of fixing those now when they will be replaced in 3.0.
Instead we just move the body declaration inside the file loop as a hotfix. Ideally, this shouldn't be necessary but the old code is hard to manage.